host/dai: update pointer cast to support 64-bit library build
This patch updates the pointer cast in host/dai for library build
support for 64-bit arch

+++ src/audio/dai.c
@@ -293,7 +293,7 @@ static int dai_playback_params(struct comp_dev *dev)
 				goto err_unwind;
 
 			elem->size = dd->period_bytes;
-			elem->src = (uint32_t)(dma_buffer->r_ptr) +
+			elem->src = (uintptr_t)(dma_buffer->r_ptr) +
 				i * dd->period_bytes;
 
 			elem->dest = dai_fifo(dd->dai, SOF_IPC_STREAM_PLAYBACK);
@@ -360,7 +360,7 @@ static int dai_capture_params(struct comp_dev *dev)
 				goto err_unwind;
 
 			elem->size = dd->period_bytes;
-			elem->dest = (uint32_t)(dma_buffer->w_ptr) +
+			elem->dest = (uintptr_t)(dma_buffer->w_ptr) +
 				i * dd->period_bytes;
 			elem->src = dai_fifo(dd->dai, SOF_IPC_STREAM_CAPTURE);
 			list_item_append(&elem->list, &config->elem_list);
